+++ mail_tracking/static/src/js/failed_message/discuss.js
@@ -1,0 +1,312 @@
+/* Copyright 2019 Alexandre Díaz
+   License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html). */
+odoo.define('mail_tracking.FailedMessageDiscuss', function (require) {
+    "use strict";
+
+    // To be considered:
+    //   - One message can be displayed in many threads
+    //   - A thread can be a mailbox, channel, ...
+    //   - A mailbox it's a type of thread that it's displayed on top of
+    //    the discuss menu, has a counter, etc...
+
+    var MailManagerNotif = require('mail.Manager.Notification');
+    var AbstractMessage = require('mail.model.AbstractMessage');
+    var Message = require('mail.model.Message');
+    var Discuss = require('mail.Discuss');
+    var MailManager = require('mail.Manager');
+    var Mailbox = require('mail.model.Mailbox');
+    var core = require('web.core');
+
+    var QWeb = core.qweb;
+    var _t = core._t;
+
+    /* The states to consider a message as failed message */
+    var FAILED_STATES = [
+        'error', 'rejected', 'spam', 'bounced', 'soft-bounced',
+    ];
+
+
+    AbstractMessage.include({
+
+        /**
+         *  Abstract declaration to know if a message its included on the
+         * failed mailbox. By default should be false.
+         *
+         * @returns {Boolean}
+         */
+        isFailed: function () {
+            return false;
+        },
+    });
+
+    Message.include({
+
+        /**
+         * Overrides to store information from server
+         *
+         * @Override
+         */
+        init: function (parent, data) {
+            this._isFailedMessage = data.is_failed_message;
+            this._super.apply(this, arguments);
+        },
+
+        /**
+        *  Implementation to know if a message its included on the
+        * failed mailbox.
+        *
+        * @Override
+        */
+        isFailed: function () {
+            return _.contains(this._threadIDs, 'mailbox_failed');
+        },
+
+        /**
+        * Adds/Remove message to/from failed mailbox
+        *
+        * @param {Boolean} failed
+        */
+        setFailed: function (failed) {
+            if (failed) {
+                this._addThread('mailbox_failed');
+            } else {
+                this.removeThread('mailbox_failed');
+            }
+        },
+
+        /**
+        *  Overrides to update if the message need be show in 'failed' mailbox
+        * thread
+        * @Override
+        */
+        _processMailboxes: function () {
+            this.setFailed(this._isFailedMessage);
+            this._super.apply(this, arguments);
+        },
+    });
+
+    MailManagerNotif.include({
+
+        /**
+         * Overrides to handle changes in the 'mail_tracking_needs_action' flag
+         *
+         * @Override
+         */
+        _handlePartnerNotification: function (data) {
+            if (data.type === 'toggle_tracking_status') {
+                this._handlePartnerToggleFailedNotification(data);
+            } else {
+                //  Workaround to avoid call '_handlePartnerChannelNotification'
+                // because this is related with the failed mailbox, not a
+                // channel.
+                this._super.apply(this, arguments);
+            }
+        },
+
+        /**
+        * This method update messages in the failed mailbox when the flag
+        * 'mail_tracking_needs_action' is toggled. This can remove/add
+        * the message from/to failed mailbox and update mailbox counter.
+        *
+        * @private
+        * @param {Object} data
+        */
+        _handlePartnerToggleFailedNotification: function (data) {
+            var self = this;
+            var failed = this.getMailbox('failed');
+            _.each(data.message_ids, function (messageID) {
+                var message = _.find(self._messages, function (msg) {
+                    return msg.getID() === messageID;
+                });
+                if (message) {
+                    message.setFailed(data.needs_actions);
+                    if (message.isFailed() === false) {
+                        self._removeMessageFromThread(
+                            'mailbox_failed', message);
+                    } else {
+                        self._addMessageToThreads(message, []);
+                        var channelFailed = self.getMailbox('failed');
+                        channelFailed.invalidateCaches();
+                    }
+                    self._mailBus.trigger('update_message', message);
+                }
+            });
+
+            if (data.needs_actions) {
+                // Increase failed counter if message is marked as failed
+                failed.incrementMailboxCounter(data.message_ids.length);
+            } else {
+                // Decrease failed counter if message is remove from failed
+                failed.decrementMailboxCounter(data.message_ids.length);
+            }
+
+            // Trigger event to refresh threads
+            this._mailBus.trigger('update_failed', failed.getMailboxCounter());
+        },
+    });
+
+    Discuss.include({
+        events: _.extend({}, Discuss.prototype.events, {
+            'click .o_failed_message_retry': '_onRetryFailedMessage',
+            'click .o_failed_message_reviewed': '_onMarkFailedMessageReviewed',
+        }),
+
+        /**
+        * Paramaters used to render 'failed' mailbox entry in Discuss
+        *
+        * @private
+        * @returns {Object}
+        */
+        _sidebarQWebParams: function () {
+            var failed = this.call('mail_service', 'getMailbox', 'failed');
+            return {
+                activeThreadID: this._thread ? this._thread.getID() : undefined,
+                failedCounter: failed.getMailboxCounter(),
+            };
+        },
+
+        /**
+        * Render 'failed' mailbox menu entry in Discuss
+        *
+        * @private
+        * @returns {jQueryElementt}
+        */
+        _renderSidebar: function () {
+            var $sidebar = this._super.apply(this, arguments);
+            // Because Odoo implementation isn't designed to be inherited
+            // properly, we inject 'failed' button using jQuery.
+            var $failed_item = $(QWeb.render('mail_tracking.SidebarFailed',
+                this._sidebarQWebParams()));
+            $failed_item.insertAfter(
+                $sidebar.find(".o_mail_discuss_title_main").filter(":last"));
+            return $sidebar;
+        },
+
+        /**
+         * Overrides to update messages in 'failed' mailbox thread
+         *
+         * @Override
+         */
+        _onMessageUpdated: function (message, type) {
+            var self = this;
+            var currentThreadID = this._thread.getID();
+            if (currentThreadID === 'mailbox_failed' && !message.isFailed()) {
+                this._thread.fetchMessages(this.domain)
+                    .then(function () {
+                        var options = self._getThreadRenderingOptions();
+                        self._threadWidget.removeMessageAndRender(
+                            message.getID(), self._thread, options)
+                            .then(function () {
+                                self._updateButtonStatus(
+                                    !self._thread.hasMessages(), type);
+                            });
+                    });
+            } else {
+                //  Workaround to avoid call '_fetchAndRenderThread' and refetch
+                // thread messages.
+                this._super.apply(this, arguments);
+            }
+        },
+
+        /**
+        *  Overrides to hide reply in the 'failed' mailbox, as this feature does
+        * not necessary in that channel. It also show retry and
+        * 'set as reviewed'.
+        *
+        * @Override
+        */
+        _getThreadRenderingOptions: function () {
+            var values = this._super.apply(this, arguments);
+            if (this._thread.getID() === 'mailbox_failed') {
+                values.displayEmailIcons = true;
+                values.displayReplyIcons = false;
+                values.displayRetryButton = true;
+                values.displayReviewedButton = true;
+            }
+            return values;
+        },
+
+        /**
+        * Overrides to listen event that refresh thread messages
+        *
+        * @Override
+        */
+        _startListening: function () {
+            this._super.apply(this, arguments);
+            this.call('mail_service', 'getMailBus')
+                .on('update_failed', this, this._throttledUpdateThreads);
+        },
+
+        // Handlers
+        /**
+         * Open the resend mail.resend.message wizard
+         *
+         * @private
+         * @param {Event} event
+         */
+        _onRetryFailedMessage: function (event) {
+            event.preventDefault();
+            var messageID = $(event.currentTarget).data('message-id');
+            this.do_action('mail.mail_resend_message_action', {
+                additional_context: {
+                    mail_message_to_resend: messageID,
+                },
+            });
+        },
+
+        /**
+        * Toggle 'mail_tracking_needs_action' flag
+        *
+        * @private
+        * @param {Event} event
+        * @returns {Promise}
+        */
+        _onMarkFailedMessageReviewed: function (event) {
+            event.preventDefault();
+            var messageID = $(event.currentTarget).data('message-id');
+            return this._rpc({
+                model: 'mail.message',
+                method: 'toggle_tracking_status',
+                args: [[messageID]],
+                context: this.getSession().user_context,
+            });
+        },
+    });
+
+    MailManager.include({
+
+        /**
+         * Overrides to add the 'failed' mailbox
+         *
+         * @Override
+         */
+        _updateMailboxesFromServer: function (data) {
+            this._super.apply(this, arguments);
+            this._addMailbox({
+                id: 'failed',
+                name: _t("Failed"),
+                mailboxCounter: data.failed_counter || 0,
+            });
+        },
+    });
+
+    Mailbox.include({
+
+        /**
+         * Overrides to add domain for 'failed' mailbox thread
+         *
+         * @Override
+         */
+        _getThreadDomain: function () {
+            if (this._id === 'mailbox_failed') {
+                return [
+                    ['mail_tracking_ids.state', 'in', FAILED_STATES],
+                    ['mail_tracking_needs_action', '=', true],
+                ];
+            }
+            // Workaround to avoid throw 'Missing domain' exception
+            return this._super.apply(this, arguments);
+        },
+    });
+
+});
