+++ mail_tracking/static/src/xml/failed_message/discuss.xml
@@ -25,15 +25,4 @@
         </t>
     </t>
 
-    <t t-extend="mail.widget.Thread.Message">
-        <t t-jquery="span[t-attf-class=o_thread_icons]" t-operation="append">
-            <a t-if="message.isFailed() &amp;&amp; options.displayRetryButton" class="btn btn-link btn-success text-success btn-sm o_failed_message_reviewed o_activity_link mr8" t-att-data-message-id="message.getID()">
-                <i class="fa fa-check"/> Set as Reviewed
-            </a>
-            <a t-if="message.isFailed() &amp;&amp; options.displayReviewedButton" class="btn btn-link btn-default text-warning btn-sm o_failed_message_retry" t-att-data-message-id="message.getID()">
-                <i class="fa fa-retweet"/> Retry
-            </a>
-        </t>
-    </t>
-
 </templates>
